Avoid crash on missing `old_dir`

+++ test_compare.py
@@ -359,7 +359,7 @@ else:
 
             for my_run in my_runs:
                 old_dir = path.join(args.compare, my_run["title"])
-                shutil.rmtree(old_dir)
+                if path.exists(old_dir): shutil.rmtree(old_dir)
                 shutil.move(my_run["title"], old_dir)
 
             dst = path.join(args.compare, "perf_report.csv")
